First working version with cards that change drawable

+++ src/org/academiadecodigo/jungleweed/HD_Grafiks_2_0/SimpleGfxCard.java
@@ -13,20 +13,18 @@ public class SimpleGfxCard extends Card {
     private Picture faceUp;
     private Picture faceDown;
 
-    private boolean hasBeenFirstDrawn;
-
     public SimpleGfxCard(CardShape shape, CardColor color) {
         super(shape, color);
-
-        faceUp = new Picture(0, 0, "res/" + shape + color + ".jpeg");
-        faceDown = new Picture(0, 0, "res/CARDBACK.jpeg");
 
     }
 
     private void recreateShapes() {
 
-        this.faceUp.delete();
-        this.faceDown.delete();
+
+        if (this.faceUp != null || this.faceDown != null) {
+            this.faceUp.delete();
+            this.faceDown.delete();
+        }
 
         faceUp = new Picture(this.getxFaceUp(), this.getyFaceUp(), "res/" + this.getShape() + this.getColor() + ".jpeg");
         faceDown = new Picture(this.getxFaceDown(), this.getyFaceDown(), "res/CARDBACK.jpeg");
@@ -42,41 +40,29 @@ public class SimpleGfxCard extends Card {
     @Override
     public void turn() {
         super.turn();
-        if (super.isVisible()) {
-            this.recreateShapes();
-        }
 
-    }
+        this.recreateShapes();
 
-    private void drawFirstTime() {
-        if (!hasBeenFirstDrawn) {
-            faceUp.draw();
-            faceDown.draw();
-        }
     }
 
     @Override
     public void setXFaceUp(int x) {
         super.setXFaceUp(x);
-        faceUp.translate(super.getxFaceUp() - x, 0);
     }
 
     @Override
     public void setYFaceUp(int y) {
         super.setYFaceUp(y);
-        faceUp.translate(0, super.getyFaceUp() - y);
     }
 
     @Override
     public void setXFaceDown(int x) {
         super.setXFaceDown(x);
-        faceDown.translate(super.getxFaceDown() - x, 0);
     }
 
     @Override
     public void setYFaceDown(int y) {
         super.setYFaceDown(y);
-        faceDown.translate(0, super.getyFaceDown() - y);
     }
 
 }
